pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "jim_bob"
version = "0.1.0"
description = "URL Health Check"
authors = [
    { name = "Joe Collins", email = "j.collins@zengenti.com" },
    { name = "Ian Glaze", email = "i.glaze@zengenti.com" },
    { name = "Nic Kilby", email = "n.kilby@zengenti.com" },
    { name = "Alex Pop", email = "a.pop@zengenti.com" }
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "dnspython",
    "flask",
    "gunicorn",
    "python-dotenv",
    "requests",
    "tinydb",
]

[project.optional-dependencies]
dev = [
    "black",
    "debugpy",
    "ipykernel",
    "isort",
    "mypy",
    "pip-tools",
    "pre-commit",
    "pytest",
    "responses",
    "ruff",
    "setuptools",
    "sourcery",
    "types-requests",  # used for type checking.
]

[project.urls]
homepage = "https://nickilby.github.io/jim-bob"
repository = "https://github.com/nickilby/jim-bob"
documentation = "https://nickilby.github.io/jim-bob"

[tool.setuptools]
py-modules = [
    "python_src",
    "tests"
]  #  Disable auto discovery of modules (and only include the ones we specifically want).

[tool.setuptools.packages.find]
where = [
    "src",
    "tests"
]

[project.scripts]
run = "jim_bob.main:main"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-q --cache-clear"
testpaths = ["./tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
follow_imports = "skip"
no_strict_optional = true
disable_error_code = "import-untyped"
ignore_missing_imports = true
